+++ assets/js/admin.js
@@ -6,52 +6,55 @@ console.log('loading admin');
 const reset_i = '<i class="icon-ccw secondary-font-color color-error-contrast"></i>'
 
 OneChat.on_load(function(one_chat) {
-  one_chat.admin = new Admin(one_chat)
-})
+  one_chat.admin = new Admin(one_chat);
+});
 
 class Admin {
   constructor(one_chat) {
-    this.one_chat = one_chat
-    this.modifed = false
-    this.register_events(this)
+    this.one_chat = one_chat;
+    this.modifed = false;
+    this.register_events(this);
   }
 
   enable_save_button() {
-    let save = $('button.save')
+    let save = $('button.save');
     if (save.attr('disabled') == 'disabled') {
-      save.removeAttr('disabled')
-      save.parent().prepend(`<button class="button danger discard"><i class="icon-send"></i><span>${gettext.cancel}</span></button>`)
-      this.modified = true
+      save.removeAttr('disabled');
+      save.parent().prepend(`<button class="button danger discard"><i class="icon-send"></i><span>${gettext.cancel}</span></button>`);
+      this.modified = true;
     }
   }
   disable_save_button() {
-    let save = $('button.save')
-    this.modified = false
-    save.attr('disabled', 'disabled')
-    $('button.discard').remove()
+    let save = $('button.save');
+    this.modified = false;
+    save.attr('disabled', 'disabled');
+    $('button.discard').remove();
   }
   add_reset_setting_button(target, input_line) {
-    let data_settings = target.getAttribute('name').replace('[', '__').replace(']', '')
-    let reset = `<button text='Reset' data-setting="${data_settings}"
-      class="reset-settings button danger" rebel-click="admin_reset_setting_click">${reset_i}</button>`
-    if (input_line.find('button.reset-settings').length === 0) {
-      input_line.append(reset);
+    let name = target.getAttribute('name');
+    if (name) {
+      let data_settings = name.replace('[', '__').replace(']', '');
+      let reset = `<button text='Reset' data-setting="${data_settings}";
+        class="reset-settings button danger" rebel-click="admin_reset_setting_click">${reset_i}</button>`;
+      if (input_line.find('button.reset-settings').length === 0) {
+        input_line.append(reset);
 
-      Rebel.set_event_handlers($(target).selector);
+        Rebel.set_event_handlers($(target).selector);
+      }
     }
   }
   handle_change(e) {
     let target = e.currentTarget;
     let input_line = $(target).closest('.input-line');
-    this.enable_save_button()
-    input_line.addClass('setting-changed')
+    this.enable_save_button();
+    input_line.addClass('setting-changed');
     this.add_reset_setting_button(target, input_line);
   }
   register_events(admin) {
     $('body')
       .on('click', 'button.discard', () => {
         // admin.disable_save_button()
-        $('a.admin-link[data-id="admin_info"]').click()
+        $('a.admin-link[data-id="admin_info"]').click();
       })
       .on('change', '.admin-settings form input:not(.search)', (e) => {
         this.handle_change(e);
@@ -66,36 +69,36 @@ class Admin {
         this.handle_change(e);
       })
       .on('change', '.permissions-manager [type="checkbox"]', function(e, t) {
-        e.preventDefault()
-        console.log('checkbox change t', $(this))
-        let name = $(this).attr('name')
-        let value = $(this).is(':checked')
+        e.preventDefault();
+        console.log('checkbox change t', $(this));
+        let name = $(this).attr('name');
+        let value = $(this).is(':checked');
 
         if (!value) { value = "false" }
         OneChat.userchan.push('admin:permissions:change:' + name, {value: value})
         .receive("ok", resp => {
           // stop_loading_animation()
-          toastr.success('Room ' + name + ' updated successfully.')
+          toastr.success('Room ' + name + ' updated successfully.');
         })
       })
       .on('click', '.page-settings .section button.expand', function(e) {
-        e.preventDefault()
+        e.preventDefault();
         $(this)
           .addClass('collapse')
           .removeClass('expand')
           .first().html('Collapse')
           .closest('.section-collapsed')
-          .removeClass('section-collapsed')
+          .removeClass('section-collapsed');
       })
       .on('click', '.page-settings .section button.collapse', function(e) {
-        e.preventDefault()
+        e.preventDefault();
         $(this)
           .removeClass('collapse')
           .addClass('expand')
           .first().html('Expand')
           .closest('.section-title')
           .parent()
-          .addClass('section-collapsed')
+          .addClass('section-collapsed');
       })
       .on('click', '.admin-settings button.save', function(e) {
         //console.log('saving form....', $('form').data('id'))
@@ -103,37 +106,37 @@ class Admin {
         OneChat.userchan.push('admin:save:' + $('form').data('id'), $('form').serializeArray())
           .receive("ok", resp => {
             if (resp.success) {
-              admin.disable_save_button()
-              toastr.success(resp.success)
+              admin.disable_save_button();
+              toastr.success(resp.success);
             } else if (resp.error) {
-              toastr.error(resp.error)
+              toastr.error(resp.error);
             }
         })
       })
       .on('click', 'button.refresh', function(e) {
-        let page = $(this).closest('section').data('page')
-        $('a.admin-link[data-id="' + page + '"]').click()
+        let page = $(this).closest('section').data('page');
+        $('a.admin-link[data-id="' + page + '"]').click();
       })
       .on('click', 'section.admin .list-view.channel-settings span[data-edit]', (e) => {
-        let channel_id = $(e.currentTarget).closest('[data-id]').data('id')
-        this.userchan_push('edit', {channel_id: channel_id, field: $(e.currentTarget).data('edit')})
+        let channel_id = $(e.currentTarget).closest('[data-id]').data('id');
+        this.userchan_push('edit', {channel_id: channel_id, field: $(e.currentTarget).data('edit')});
       })
       .on('click', 'section.admin .channel-settings button.save', e => {
-        let channel_id = $(e.currentTarget).closest('[data-id]').data('id')
-        let params = $('.channel-settings form').serializeArray()
-        this.userchan_push('save', {channel_id: channel_id, params: params})
+        let channel_id = $(e.currentTarget).closest('[data-id]').data('id');
+        let params = $('.channel-settings form').serializeArray();
+        this.userchan_push('save', {channel_id: channel_id, params: params});
       })
       .on('click', 'section.admin .channel-settings button.cancel', e => {
-        let channel_id = $(e.currentTarget).closest('[data-id]').data('id')
-        this.userchan_push('cancel', {channel_id: channel_id})
+        let channel_id = $(e.currentTarget).closest('[data-id]').data('id');
+        this.userchan_push('cancel', {channel_id: channel_id});
       })
       .on('click', '#showPassword', e => {
         let prefix = "";
         if ($('#user_password').length > 0) {
           prefix = "user_";
         }
-        let password_name = `#${prefix}password`
-        let password_confirmation_name = `#${prefix}password_confirmation`
+        let password_name = `#${prefix}password`;
+        let password_confirmation_name = `#${prefix}password_confirmation`;
         $(e.currentTarget).hide();
         $('#hidePassword').show();
         $(password_name).attr('type','text');
@@ -144,26 +147,26 @@ class Admin {
         if ($('#user_password').length > 0) {
           prefix = "user_";
         }
-        let password_name = `#${prefix}password`
-        let password_confirmation_name = `#${prefix}password_confirmation`
+        let password_name = `#${prefix}password`;
+        let password_confirmation_name = `#${prefix}password_confirmation`;
         $(e.currentTarget).hide();
         $('#showPassword').show();
         $(password_name).attr('type','password');
         $(password_confirmation_name).attr('type','password');
       })
       .on('click', '#randomPassword', e => {
-        let new_password = OneChat.randomString(12)
+        let new_password = OneChat.randomString(12);
         let prefix = "";
         if ($('#user_password').length > 0) {
           prefix = "user_";
         }
-        let password_name = `#${prefix}password`
-        let password_confirmation_name = `#${prefix}password_confirmation`
+        let password_name = `#${prefix}password`;
+        let password_confirmation_name = `#${prefix}password_confirmation`;
 
-        e.preventDefault()
-        e.stopPropagation()
-        $(password_name).attr('type', 'password').val(new_password)
-        $(password_confirmation_name).attr('type', 'password').val(new_password)
+        e.preventDefault();
+        e.stopPropagation();
+        $(password_name).attr('type', 'password').val(new_password);
+        $(password_confirmation_name).attr('type', 'password').val(new_password);
         $('#showPassword').show();
         $('#hidePassword').hide();
       })
@@ -171,36 +174,35 @@ class Admin {
         OneChat.userchan.push('admin:save:user', $('form.user').serializeArray())
           .receive("ok", resp => {
             if (resp.success) {
-              toastr.success(resp.success)
-              this.close_edit_form($('form.user').data('username'))
+              toastr.success(resp.success);
+              this.close_edit_form($('form.user').data('username'));
             } else if (resp.error) {
-              toastr.error(resp.error)
+              toastr.error(resp.error);
             }
           })
           .receive("error", resp => {
-            console.log('error resp', resp)
+            console.log('error resp', resp);
             if (resp.error) {
-              toastr.error(resp.error)
+              toastr.error(resp.error);
             }
             if (resp.errors) {
-              this.show_form_errors(resp.errors)
+              this.show_form_errors(resp.errors);
             }
           })
       })
       .on('click', 'section.admin form.user button.cancel', e => {
-        this.close_edit_form($('form.user').data('username'))
+        this.close_edit_form($('form.user').data('username'));
       })
       .on('click', 'a.new-role', e => {
-        OneChat.userchan.push('admin:permissions:role:new', {})
+        OneChat.userchan.push('admin:permissions:role:new', {});
       })
       .on('click', 'a[href="#admin-permissions-edit"]', e => {
-        let name = $(e.currentTarget).attr('name')
-        console.log('permissions edit', name)
-        OneChat.userchan.push('admin:permissions:role:edit', {name: name})
+        let name = $(e.currentTarget).attr('name');
+        OneChat.userchan.push('admin:permissions:role:edit', {name: name});
       })
       .on('click', '.admin-role.delete', e => {
-        let name = $(e.currentTarget).attr('data-name')
-        OneChat.userchan.push('admin:permissions:role:delete', {name: name})
+        let name = $(e.currentTarget).attr('data-name');
+        OneChat.userchan.push('admin:permissions:role:delete', {name: name});
       })
       .on('click', 'a[href="/admin/permissions"]', e => {
         e.preventDefault();