+++ assets/js/chat_dropzone.js
@@ -6,40 +6,47 @@ $(document).ready(() => {
     return (('draggable' in div) || ('ondragstart' in div && 'ondrop' in div)) && 'FormData' in window && 'FileReader' in window;
   }();
   if (isAdvancedUpload) {
-    let droppedFiles = false
+    let droppedFiles = false;
     let enterTarget = null;
-    let obj = $('.dropzone')
+    let obj = $('.dropzone');
 
     $('body').on('drag dragstart dragend dragover dragenter dragleave drop', '.dropzone', e => {
-      e.preventDefault()
-      e.stopPropagation()
+      e.preventDefault();
+      e.stopPropagation();
     })
     .on('dragover dragenter', '.dropzone', (e) => {
       if (chat_settings.allow_upload) {
-        enterTarget = e.target
-        $('.dropzone').addClass('over')
+        enterTarget = e.target;
+        $('.dropzone').addClass('over');
       }
     })
     .on('dragleave dragend drop', '.dropzone', (e) => {
       if (enterTarget == e.target) {
-        $('.dropzone').removeClass('over')
+        $('.dropzone').removeClass('over');
       }
     })
     .on('drop', '.avatar-dropzone', event => {
       // handle the avatar image file drop
       if (chat_settings.allow_upload) {
-        let e = event.originalEvent || event
-        let files = e.dataTransfer.files || []
-        OneChat.fileUpload.handleAvatarUpload(files, obj)
+        let e = event.originalEvent || event;
+        let files = e.dataTransfer.files || [];
+        OneChat.fileUpload.handleAvatarUpload(files, obj);
       }
     })
-    .on('drop', '.dropzone:not(.avatar-dropzone)', event => {
+    .on('drop', '.site-avatar-dropzone', event => {
+      if (chat_settings.allow_upload) {
+        let e = event.originalEvent || event;
+        let files = e.dataTransfer.files || [];
+        OneChat.fileUpload.handleSiteAvatarUpload(files, obj);
+      }
+    })
+    .on('drop', '.attachment-dropzone', event => {
       // handle the attachment file drop
       if (chat_settings.allow_upload) {
-        let e = event.originalEvent || event
-        let files = e.dataTransfer.files || []
-        OneChat.fileUpload.handleFileUpload(files, obj)
+        let e = event.originalEvent || event;
+        let files = e.dataTransfer.files || [];
+        OneChat.fileUpload.handleFileUpload(files, obj);
       }
-    })
+    });
   }
-})
+});