+++ priv/repo/migrations/20180330014650_add_site_avatar_to_settings_general.exs
@@ -1,0 +1,10 @@
+defmodule InfinityOnePages.Repo.Migrations.AddSiteAvatarToSettingsGeneral do
+  use Ecto.Migration
+
+  def change do
+    alter table(:settings_general) do
+      add :site_avatar, :string
+      add :site_client_name, :string, default: "use-host-name"
+    end
+  end
+end